move ownbeaconstatus to write dir
