+++ src/components/views/beacon/OwnBeaconStatus.tsx
@@ -19,8 +19,8 @@ import React, { HTMLProps } from 'react';
 
 import { _t } from '../../../languageHandler';
 import { useOwnLiveBeacons } from '../../../utils/beacon';
-import BeaconStatus from '../beacon/BeaconStatus';
-import { BeaconDisplayStatus } from '../beacon/displayStatus';
+import BeaconStatus from './BeaconStatus';
+import { BeaconDisplayStatus } from './displayStatus';
 import AccessibleButton from '../elements/AccessibleButton';
 
 interface Props {
@@ -56,33 +56,33 @@ const OwnBeaconStatus: React.FC<Props & HTMLProps<HTMLDivElement>> = ({
         displayLiveTimeRemaining
         {...rest}
     >
-        { ownDisplayStatus === BeaconDisplayStatus.Active && <AccessibleButton
+        {ownDisplayStatus === BeaconDisplayStatus.Active && <AccessibleButton
             data-test-id='beacon-status-stop-beacon'
             kind='link'
             onClick={onStopSharing}
             className='mx_BeaconStatus_destructiveButton'
             disabled={stoppingInProgress}
         >
-            { _t('Stop') }
+            {_t('Stop')}
         </AccessibleButton>
         }
-        { hasWireError && <AccessibleButton
+        {hasWireError && <AccessibleButton
             data-test-id='beacon-status-reset-wire-error'
             kind='link'
             onClick={onResetWireError}
             className='mx_BeaconStatus_destructiveButton'
         >
-            { _t('Retry') }
+            {_t('Retry')}
         </AccessibleButton>
         }
-        { hasStopSharingError && <AccessibleButton
+        {hasStopSharingError && <AccessibleButton
             data-test-id='beacon-status-stop-beacon-retry'
             kind='link'
             onClick={onStopSharing}
             className='mx_BeaconStatus_destructiveButton'
         >
-            { _t('Retry') }
-        </AccessibleButton> }
+            {_t('Retry')}
+        </AccessibleButton>}
     </BeaconStatus>;
 };
